Bunch of JEI integration improvements.
NEI's item panel is back, can switch between it and JEI's
If JEI's item panel is up, it will default to JEI recipes and nei's will
be the fall back. This is reversed if NEI's panel is up.

TODO:
Subsets still need a little bit of work.
JEI and NEI need to respect each others hidden items.